+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for Lovely Webservice Status
 unreleased
 ==========
 
+ - provide metrics from prometheus_client if the packages is installed
+
 2018/04/17 0.0.5
 ================